[MDT]: Write out a basic roadmap for the series
This commit marks the start of what I call the Modernize Django
Tooling Series, where I take a Django application written a
few years ago and improve it using modern tooling.

The django application in question is just a simple URL shortener,
just enough to be a Django application, it does not use any of
Django's great features at all.